// x-pack/platform/packages/shared/response-ops/alerts-table/types.d.ts
export type CasesOwner = 'securitySolution' | 'observability' | 'cases';
// ...
export interface AlertsTableSortCombinations {
    [field: string]: {
        order: 'asc' | 'desc';
    };
}
export interface AlertsTableProps<AC extends AdditionalContext = AdditionalContext> extends PublicAlertsDataGridProps {
    /**
     * A unique identifier used to persist the table state in localStorage
     */
    id: string;
    /**
     * The columns to be displayed in the table (row selection checkboxes
     * and actions column are prepended automatically)
     */
    columns?: EuiDataGridProps['columns'];
// ...
}
/**
 * A utility type to extract the type of a prop from `AlertsTableProps`, excluding `undefined`.
 */
export type GetAlertsTableProp<PropKey extends keyof AlertsTableProps> = NonNullable<AlertsTableProps[PropKey]>;
export interface AlertsTableImperativeApi {
    refresh: () => void;
    toggleColumn: (columnId: string) => void;
}
export type AlertsTablePropsWithRef<AC extends AdditionalContext> = AlertsTableProps<AC> & RefAttributes<AlertsTableImperativeApi>;
export interface AdditionalContext {
}
export type RenderContext<AC extends AdditionalContext> = {
    tableId?: string;
    dataGridRef: MutableRefObject<EuiDataGridRefProps | null>;
    /**
     * Refetches all the queries, resetting the alerts pagination if necessary
     */
    refresh: () => void;
    /**
     * True if any of the active queries is fetching
     */
    isLoading: boolean;
    isLoadingAlerts: boolean;
    alerts: Alert[];
    alertsCount: number;
    browserFields: BrowserFields;
    isLoadingMutedAlerts: boolean;
    mutedAlerts?: MutedAlerts;
    isLoadingCases: boolean;
    cases?: Map<string, Case>;
    isLoadingMaintenanceWindows: boolean;
    maintenanceWindows?: Map<string, MaintenanceWindow>;
    showAlertStatusWithFlapping?: boolean;
    bulkActionsStore: [BulkActionsState, Dispatch<BulkActionsReducerAction>];
} & SetRequired<Pick<AlertsTableProps<AC>, 'columns' | 'pageIndex' | 'onPageIndexChange' | 'pageSize' | 'onPageSizeChange' | 'renderCellValue' | 'renderCellPopover' | 'renderActionsCell' | 'expandedAlertIndex' | 'onExpandedAlertIndexChange' | 'renderExpandedAlertView' | 'services' | 'casesConfiguration' | 'openLinksInNewTab' | 'isMutedAlertsEnabled' | 'getAlertFormatter' | 'alertDetailsNavigation'>, 'columns' | 'pageIndex' | 'onPageIndexChange' | 'pageSize' | 'onPageSizeChange' | 'openLinksInNewTab' | 'expandedAlertIndex' | 'onExpandedAlertIndexChange' | 'isMutedAlertsEnabled'> & AC;
export type ComponentRenderer<AC extends AdditionalContext> = ComponentType<RenderContext<AC>>;
export interface CellActionsOptions {
    /**
     * Resolves the cell actions for a given column
     */
    getCellActionsForColumn: (columnId: string, columnIndex: number) => EuiDataGridColumnCellAction[];
    visibleCellActions?: number;
    disabledCellActions?: string[];
}
export interface PublicAlertsDataGridProps extends Omit<EuiDataGridProps, 'renderCellPopover' | 'renderCellValue' | 'aria-labelledby' | 'columnVisibility' | 'rowCount' | 'sorting' | 'cellContext' | 'pagination' | 'columns'> {
    ruleTypeIds: string[];
    minScore?: number;
    trackScores?: boolean;
    consumers?: string[];
    /**
     * Value to override the server side search
     */
    projectRouting?: ProjectRouting;
    /**
     * If true, shows a button in the table toolbar to inspect the search alerts request
     */
    showInspectButton?: boolean;
    /**
     * Cases-specific configuration options
     */
    casesConfiguration?: {
        featureId: string;
        owner: CasesOwner[];
        appId?: string;
        syncAlerts?: boolean;
        extractObservables?: boolean;
    };
    /**
     * If true, hides the bulk actions controls
     */
    hideBulkActions?: boolean;
    /**
     * An array of bulk actions to be displayed in the table
     */
    additionalBulkActions?: BulkActionsPanelConfig[];
    /**
     * Width of the actions column
     */
    actionsColumnWidth?: number;
    /**
     * Options passed to the fields browser modal
     */
    fieldsBrowserOptions?: FieldBrowserOptions;
    /**
     * Options to customize the actions menu for each cell
     */
    cellActionsOptions?: CellActionsOptions;
}
export interface AlertsDataGridProps<AC extends AdditionalContext = AdditionalContext> extends PublicAlertsDataGridProps, Pick<EuiDataGridProps, 'columnVisibility'> {
    renderContext: RenderContext<AC>;
    additionalToolbarControls?: ReactNode;
    pageSizeOptions?: number[];
    leadingControlColumns?: EuiDataGridControlColumn[];
    trailingControlColumns?: EuiDataGridControlColumn[];
    'data-test-subj': string;
    onToggleColumn: (columnId: string) => void;
    onResetColumns: () => void;
    onColumnResize?: EuiDataGridOnColumnResizeHandler;
    query: Pick<NonNullable<QueryDslQueryContainer>, 'bool' | 'ids'>;
    showInspectButton?: boolean;
    toolbarVisibility?: EuiDataGridToolBarVisibilityOptions;
    /**
     * Allows to consumers of the table to decide to highlight a row based on the current alert.
     */
    shouldHighlightRow?: (alert: Alert) => boolean;
    /**
     * Enable when rows may have variable heights (disables virtualization)
     */
    dynamicRowHeight?: boolean;
    sort: AlertsTableSortCombinations[];
    onSortChange: (sort: EuiDataGridSorting['columns']) => void;
    alertsQuerySnapshot?: EsQuerySnapshot;
}
// ...
export type BulkActionsPanelConfig = ItemsPanelConfig | ContentPanelConfig;
export declare enum BulkActionsVerbs {
    add = "add",
    delete = "delete",
    clear = "clear",
    selectCurrentPage = "selectCurrentPage",
    selectAll = "selectAll",
    rowCountUpdate = "rowCountUpdate",
    updateRowLoadingState = "updateRowLoadingState",
    updateAllLoadingState = "updateAllLoadingState"
}
export interface BulkActionsReducerAction {
    action: BulkActionsVerbs;
    rowIndex?: number;
    rowCount?: number;
    isLoading?: boolean;
}
export interface BulkActionsState {
    rowSelection: Map<number, RowSelectionState>;
    isAllSelected: boolean;
    areAllVisibleRowsSelected: boolean;
    rowCount: number;
    updatedAt: number;
}
export interface BulkEditTagsFlyoutState {
    isFlyoutOpen: boolean;
    onClose: () => void;
    onSaveTags: (itemsSelection: ItemsSelectionState) => void;
}
export interface AlertsTableFlyoutBaseProps {
    alert: Alert;
    isLoading: boolean;
    id?: string;
}
export type RowSelection = Map<number, RowSelectionState>;
export interface RowSelectionState {
    isLoading: boolean;
}